requirements.txt
Django==5.1
docker==7.1.0
requests==2.32.3
gunicorn
whitenoise
django-q2==1.8.0
django-hosts==7.0.0
psycopg2-binary==2.9.11
